fix: require babel-register from correct directory
Problem: invoking `product-space` command on the command line caused an
error because `node` cannot parse ES modules.

Reason: Because we're using `babel-register`, it shouldn't matter that
ES module is used. The real reason for the error is that
`require(babel-register)` is invoked before the directory has been
changed to the directory of the `bin.js` script. Probably the effect is
that `require('babel-register')` doesn't work in that case.

Solution: `cd` into the directory of the project (where we are sure that
`babel-register` has been installed) before invoking `babel-register`.