Change ID in built-in data sets to HS/SITC codes
Previously, the IDs in the sample HS/SITC network layout/metadata files
are the internal IDs used in the international atlas. However, this is
inconvenient for the users because they don't (and shouldn't have to)
know about the relationship between these IDs and the offical 4-digit
codes. This commit replace those internal IDs with the official codes.