Allow more flexible data input format
Previously, the format of the input data was quite strict e.g. `values`
has to be an array of numbers and `active` has to be a boolean. This was
problematic because Stata is not capable of outputting JSON files where
the values of keys can be an array and Stata doesn't have built-in
boolean values (just 0 or 1). This commit takes accommodates these
limitations to make it easier for the user to input data.