Vertically flip orientation of network layout
Previously, the network's nodes' coordinates are mapped such that larger
y values appear towards the top and smaller ones towards the bottom,
which is how normal coords work. However, this causes the standard
layout for HS and SITC product space to appear vertically flipped
compared to the (probably wrong) rendered layout of the product HS/SITC
product space in the international atlas (old and new). This commit
reverses the current approach so that the rendered layout matches that
in the international atlas.